refactor(core): address PR #33 review feedback
