+++ src/chatboteval/core/schemas/__init__.py
@@ -1,24 +1,1 @@
-"""Public re-export surface for core.schemas."""
-
-from chatboteval.core.schemas.annotation_export import (
-    AnnotationBase,
-    GenerationAnnotation,
-    GroundingAnnotation,
-    RetrievalAnnotation,
-)
-from chatboteval.core.schemas.annotation_import import Chunk, QueryResponsePair
-from chatboteval.core.schemas.base import ContractModel, Task
-from chatboteval.core.schemas.csv_io import read_csv, write_csv
-
-__all__ = [
-    "Task",
-    "ContractModel",
-    "Chunk",
-    "QueryResponsePair",
-    "AnnotationBase",
-    "RetrievalAnnotation",
-    "GroundingAnnotation",
-    "GenerationAnnotation",
-    "write_csv",
-    "read_csv",
-]
+"""Core boundary contracts."""
